third_party/blink/web_tests/platform/generic/fast/forms/radio/ValidityState-valueMissing-radio-expected.txt

valueMissing tests for radio buttons

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Without form element
No checked button:
PASS inputs[0].validity.valueMissing is true
PASS inputs[1].validity.valueMissing is true
PASS inputs[2].validity.valueMissing is true
The second button has been checked:
PASS inputs[0].validity.valueMissing is false
PASS inputs[1].validity.valueMissing is false
PASS inputs[2].validity.valueMissing is false
The first button has been checked:
PASS inputs[0].validity.valueMissing is false
PASS inputs[1].validity.valueMissing is false
PASS inputs[2].validity.valueMissing is false
The third button has been checked:
PASS inputs[0].validity.valueMissing is false
PASS inputs[1].validity.valueMissing is false
PASS inputs[2].validity.valueMissing is false


With form element
No checked button:
PASS inputs[0].validity.valueMissing is true
PASS inputs[1].validity.valueMissing is true
PASS inputs[2].validity.valueMissing is true
The first button has been checked:
PASS inputs[0].validity.valueMissing is false
PASS inputs[1].validity.valueMissing is false
PASS inputs[2].validity.valueMissing is false
The second button has been checked:
PASS inputs[0].validity.valueMissing is false
PASS inputs[1].validity.valueMissing is false
PASS inputs[2].validity.valueMissing is false
The third button has been checked:
PASS inputs[0].validity.valueMissing is false
PASS inputs[1].validity.valueMissing is false
PASS inputs[2].validity.valueMissing is false
PASS successfullyParsed is true

TEST COMPLETE
